Security context features extracted into extra context.
